Providers: custom providers supported. Leaflet plugin deleted
