+++ src/index.js
@@ -1,7 +1,6 @@
 export { default } from './layer-manager';
 
 // Plugins
-// export { default as PluginLeaflet } from './plugins/plugin-leaflet';
 export { default as PluginMapboxGl } from './plugins/plugin-mapbox-gl';
 
 // Utils
