+++ src/index.umd.js
@@ -3,7 +3,6 @@
 // compatible with mixed imports (which exist in index.js)
 // This file does away with named imports in favor of a single export default.
 
-// import PluginLeaflet from './plugins/plugin-leaflet';
 import PluginMapboxGl from './plugins/plugin-mapbox-gl';
 
 import { replace, substitution, concatenation } from './utils/query';
