+++ src/plugins/plugin-mapbox-gl/video-layer-mapbox-gl.js
@@ -1,25 +1,13 @@
 import Promise from 'utils/promise';
 
-import { replace } from 'utils/query';
-
 const VideoLayer = layerModel => {
-  const { source = {}, render = {}, params, sqlParams, id } = layerModel;
-
-  const sourceParsed =
-    source.parse === false
-      ? source
-      : JSON.parse(replace(JSON.stringify(source), params, sqlParams));
-
-  const renderParsed =
-    render.parse === false
-      ? source
-      : JSON.parse(replace(JSON.stringify(render), params, sqlParams));
+  const { source = {}, render, id } = layerModel;
 
   const layer = {
     id,
     type: 'raster',
     source: {
-      ...sourceParsed,
+      ...source,
       type: 'video-tiled',
       // tiles: [ "https://storage.googleapis.com/skydipper_materials/movie-tiles/MODIS/{z}/{x}/{y}.mp4" ],
       tiles: [
@@ -32,7 +20,7 @@ const VideoLayer = layerModel => {
       playbackRate: 0.1
       // geometryFilter: 'https://storage.googleapis.com/deltares-video-map/mapbox-test/test1/geometry-filter.geojson'
     },
-    ...renderParsed,
+    ...render,
     layers: [
       {
         id: `${id}-raster`,
