--- src/services/carto-service.js
@@ -1,57 +1,0 @@
-import { CancelToken } from 'axios';
-import { get } from 'lib/request';
-import { replace } from 'utils/query';
-
-export const fetchCartoAnonymous = layerModel => {
-  const { source, params, sqlParams, interactivity } = layerModel;
-
-  const sourceParsed =
-    source.parse === false
-      ? source
-      : JSON.parse(replace(JSON.stringify(source), params, sqlParams));
-
-  const { provider } = sourceParsed;
-
-  const layerTpl = JSON.stringify({
-    version: '1.3.0',
-    stat_tag: 'API',
-    layers: provider.options.layers.map(l => {
-      if (!!interactivity && interactivity.length) {
-        return { ...l, options: { ...l.options, interactivity } };
-      }
-      return l;
-    })
-  });
-
-  // https://carto.com/developers/auth-api/guides/how-to-send-API-Keys/
-  const apiParams = {
-    stat_tag: 'API',
-    config: encodeURIComponent(layerTpl),
-    ...(provider.options.api_key && { api_key: provider.options.api_key })
-  };
-  const apiParamsString = Object.keys(apiParams)
-    .map(k => `${k}=${apiParams[k]}`)
-    .join('&');
-  const url = `https://${provider.options.account}.carto.com/api/v1/map?${apiParamsString}`;
-
-  const { layerRequest } = layerModel;
-  if (layerRequest) {
-    layerRequest.cancel('Operation canceled by the user.');
-  }
-
-  const layerRequestSource = CancelToken.source();
-  layerModel.set('layerRequest', layerRequestSource);
-
-  const newLayerRequest = get(url, { cancelToken: layerRequestSource.token }).then(res => {
-    if (res.status > 400) {
-      console.error(res);
-      return false;
-    }
-
-    return res.data;
-  });
-
-  return newLayerRequest;
-};
-
-export default { fetchCartoAnonymous };
